+++ src/convert/mod.rs
@@ -58,9 +58,10 @@ use core::iter::FusedIterator;
 pub fn validation_to_result<T, E>(validation: Validation<E, T>) -> Result<T, E> {
     match validation {
         Validation::Valid(value) => Ok(value),
-        Validation::Invalid(mut errors) => {
+        Validation::Invalid(errors) => {
             let error = errors
-                .pop()
+                .into_iter()
+                .next()
                 .expect("Validation::Invalid must contain at least one error");
             Err(error)
         },